+++ src/components/MultiSelect/MultiSelect.stories.tsx
@@ -212,6 +212,9 @@ const meta: Meta<typeof MultiSelect> = {
     id: {
       control: false,
     },
+    closeOnBlur: {
+      table: { defaultValue: { summary: false } },
+    },
     isBlockElement: {
       table: { defaultValue: { summary: false } },
     },
@@ -250,6 +253,7 @@ export const withControls: Story = {
   args: {
     buttonText: "MultiSelect",
     id: "multi-select-id",
+    closeOnBlur: true,
     isBlockElement: true,
     isDefaultOpen: false,
     isSearchable: false,
@@ -467,8 +471,19 @@ export const defaultOpenState: Story = {
   ),
 };
 
+export const closeOnBlurState: Story = {
+  render: () => (
+    <MultiSelectStory
+      id="multi-select-id-15"
+      closeOnBlur={true}
+      isBlockElement
+      items={withChildrenItems}
+    />
+  ),
+};
+
 export const InAGroup: Story = {
-  render: () => <MultiSelecGroupStory items={withItems} />,
+  render: () => <MultiSelectGroupStory items={withItems} />,
 };
 
 const MultiSelectWithControlsStory = (args) => {
@@ -511,6 +526,7 @@ const MultiSelectWithControlsStory = (args) => {
 
 const MultiSelectStory = ({
   id,
+  closeOnBlur = false,
   isBlockElement = false,
   isSearchable = false,
   isDefaultOpen,
@@ -538,6 +554,7 @@ const MultiSelectStory = ({
   return (
     <MultiSelect
       buttonText="MultiSelect"
+      closeOnBlur={closeOnBlur}
       defaultItemsVisible={defaultItemsVisible}
       id={id}
       isBlockElement={isBlockElement}
@@ -564,7 +581,7 @@ const MultiSelectStory = ({
 };
 
 // TODO: Replace with MultiSelectGroup once that component is done.
-const MultiSelecGroupStory = ({ items }: Partial<MultiSelectProps>) => {
+const MultiSelectGroupStory = ({ items }: Partial<MultiSelectProps>) => {
   // Example with custom hook useMultiSelect.
   const { onChange, onMixedStateChange, onClear, onClearAll, selectedItems } =
     useMultiSelect();